Add quarkus test profiles for h2 and PostgresSql
