+++ data-index/data-index-service/data-index-service-infinispan/pom.xml
@@ -97,6 +97,10 @@
       <scope>test</scope>
     </dependency>
     <dependency>
+      <groupId>org.infinispan</groupId>
+      <artifactId>infinispan-client-hotrod</artifactId>
+    </dependency>
+    <dependency>
       <groupId>org.kie.kogito</groupId>
       <artifactId>kogito-quarkus-test-utils</artifactId>
       <scope>test</scope>
