+++ jobs-service/jobs-service-storage-jpa/pom.xml
@@ -79,6 +79,11 @@
             <scope>test</scope>
         </dependency>
         <dependency>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-jdbc-postgresql</artifactId>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
             <groupId>org.kie.kogito</groupId>
             <artifactId>kogito-quarkus-test-utils</artifactId>
             <scope>test</scope>
