+++ jobs-service/jobs-service-storage-jpa/src/test/java/org/kie/kogito/jobs/service/repository/jpa/PostgresSqlJPAReactiveJobRepositoryTest.java
@@ -16,18 +16,36 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.jobs.service.resource;
+package org.kie.kogito.jobs.service.repository.jpa;
 
-import org.kie.kogito.jobs.service.profiles.H2QuarkusTestProfile;
+import org.junit.jupiter.api.BeforeEach;
+import org.kie.kogito.jobs.service.profiles.PostgreSQLQuarkusTestProfile;
+import org.kie.kogito.jobs.service.repository.ReactiveJobRepository;
+import org.kie.kogito.jobs.service.repository.impl.BaseJobRepositoryTest;
+import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.h2.H2DatabaseTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
-@QuarkusTest
-@QuarkusTestResource(value = H2DatabaseTestResource.class, restrictToAnnotatedClass = true)
-@TestProfile(H2QuarkusTestProfile.class)
-public class JPAJobResourceTest extends BaseJobResourceTest {
+import jakarta.inject.Inject;
 
+@QuarkusTest
+@QuarkusTestResource(value = PostgreSqlQuarkusTestResource.class, restrictToAnnotatedClass = true)
+@TestProfile(PostgreSQLQuarkusTestProfile.class)
+public class PostgresSqlJPAReactiveJobRepositoryTest extends BaseJobRepositoryTest {
+
+    @Inject
+    JPAReactiveJobRepository tested;
+
+    @BeforeEach
+    public void setUp() throws Exception {
+
+        super.setUp();
+    }
+
+    @Override
+    public ReactiveJobRepository tested() {
+        return tested;
+    }
 }
