+++ jobs-service/jobs-service-storage-jpa/src/test/java/org/kie/kogito/jobs/service/repository/jpa/PostgresSqlJPAReactiveJobServiceManagementRepositoryTest.java
@@ -16,18 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.jobs.service.resource;
+package org.kie.kogito.jobs.service.repository.jpa;
 
-import org.kie.kogito.jobs.service.profiles.H2QuarkusTestProfile;
+import org.kie.kogito.jobs.service.profiles.PostgreSQLQuarkusTestProfile;
+import org.kie.kogito.testcontainers.quarkus.PostgreSqlQuarkusTestResource;
 
 import io.quarkus.test.common.QuarkusTestResource;
-import io.quarkus.test.h2.H2DatabaseTestResource;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
 
 @QuarkusTest
-@QuarkusTestResource(value = H2DatabaseTestResource.class, restrictToAnnotatedClass = true)
-@TestProfile(H2QuarkusTestProfile.class)
-public class JPAJobResourceTest extends BaseJobResourceTest {
+@QuarkusTestResource(value = PostgreSqlQuarkusTestResource.class, restrictToAnnotatedClass = true)
+@TestProfile(PostgreSQLQuarkusTestProfile.class)
+class PostgresSqlJPAReactiveJobServiceManagementRepositoryTest extends BaseJPAReactiveJobServiceManagementRepositoryTest {
 
 }
