+++ jobs-service/jobs-service-storage-jpa/src/test/resources/application.properties
@@ -19,9 +19,11 @@
 # Kogito
 kogito.apps.persistence.type=jdbc
 # Data source
-quarkus.datasource.db-kind=h2
-quarkus.datasource.username=kogito
-quarkus.datasource.jdbc.url=jdbc:h2:mem:default;NON_KEYWORDS=VALUE,KEY
+%test-h2.quarkus.datasource.db-kind=h2
+%test-h2.quarkus.datasource.username=kogito
+%test-h2.quarkus.datasource.jdbc.url=jdbc:h2:mem:default;NON_KEYWORDS=VALUE,KEY
+%test-postgresql.quarkus.datasource.db-kind=postgresql
+%test-postgresql.quarkus.datasource.devservices.enabled=false
 kie.flyway.enabled=true
 # Disabling Security for tests
 quarkus.oidc.enabled=false
