+++ persistence-commons/persistence-commons-infinispan/pom.xml
@@ -49,6 +49,10 @@
       <groupId>io.quarkus</groupId>
       <artifactId>quarkus-infinispan-client</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.infinispan</groupId>
+      <artifactId>infinispan-client-hotrod</artifactId>
+    </dependency>
 
     <!-- Health Check -->
     <dependency>
